backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.0.1
opencv-python>=4.9.0
torch>=2.0.0
torchvision>=0.15.0
numpy<2.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
scikit-learn>=1.3.0
matplotlib>=3.7.2
pydantic==2.4.2
pydantic-settings==2.1.0
sqlalchemy==2.0.23
httpx==0.25.0